Implement realtime within v1 API
