+++ pkg/api/apiv1/apiv1.go
@@ -1,7 +1,6 @@
 package apiv1
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"net/http"
@@ -9,10 +8,12 @@ import (
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
-	"github.com/google/uuid"
+	"github.com/inngest/inngest/pkg/api/apiv1/apiv1auth"
+	"github.com/inngest/inngest/pkg/consts"
 	"github.com/inngest/inngest/pkg/cqrs"
 	"github.com/inngest/inngest/pkg/execution"
 	"github.com/inngest/inngest/pkg/execution/queue"
+	"github.com/inngest/inngest/pkg/execution/realtime"
 	"github.com/inngest/inngest/pkg/execution/state/redis_state"
 	"github.com/inngest/inngest/pkg/headers"
 )
@@ -25,7 +26,7 @@ type Opts struct {
 	// a max-age.
 	CachingMiddleware CachingMiddleware
 	// WorkspaceFinder returns the authenticated workspace given the current context.
-	AuthFinder AuthFinder
+	AuthFinder apiv1auth.AuthFinder
 	// Executor is required to cancel and manage function executions.
 	Executor execution.Executor
 	// EventReader allows reading of events from storage.
@@ -40,12 +41,14 @@ type Opts struct {
 	CancellationReadWriter cqrs.CancellationReadWriter
 	// QueueShardSelector determines the queue shard to use
 	QueueShardSelector redis_state.ShardSelector
+	// Broadcaster is used to handle realtime via APIv1
+	Broadcaster realtime.Broadcaster
 }
 
 // AddRoutes adds a new API handler to the given router.
 func AddRoutes(r chi.Router, o Opts) http.Handler {
 	if o.AuthFinder == nil {
-		o.AuthFinder = NilAuthFinder
+		o.AuthFinder = apiv1auth.NilAuthFinder
 	}
 
 	// Create the HTTP implementation, which wraps the handler.  We do ths to code
@@ -82,6 +85,14 @@ func (a *router) setup() {
 		}
 
 		r.Use(headers.ContentTypeJsonResponse())
+
+		rt := realtime.NewAPI(realtime.APIOpts{
+			JWTSecret:      consts.DevServerRealtimeJWTSecret,
+			Broadcaster:    a.opts.Broadcaster,
+			AuthMiddleware: a.opts.AuthMiddleware,
+			AuthFinder:     a.opts.AuthFinder,
+		})
+		r.Mount("/", rt)
 
 		r.Get("/events", a.getEvents)
 		r.Get("/events/{eventID}", a.getEvent)
@@ -131,30 +142,4 @@ type Response[T any] struct {
 type ResponseMetadata struct {
 	FetchedAt   time.Time  `json:"fetched_at,omitempty"`
 	CachedUntil *time.Time `json:"cached_until,omitempty"`
-}
-
-// TODO (tonyhb) Open source the auth context.
-
-// AuthFinder returns auth information from the current context.
-type AuthFinder func(ctx context.Context) (V1Auth, error)
-
-// V1Auth represents an object that returns the account and worskpace currently authed.
-type V1Auth interface {
-	AccountID() uuid.UUID
-	WorkspaceID() uuid.UUID
-}
-
-// NilAuthFinder is used in the dev server, returning zero auth.
-func NilAuthFinder(ctx context.Context) (V1Auth, error) {
-	return nilAuth{}, nil
-}
-
-type nilAuth struct{}
-
-func (nilAuth) AccountID() uuid.UUID {
-	return uuid.UUID{}
-}
-
-func (nilAuth) WorkspaceID() uuid.UUID {
-	return uuid.UUID{}
 }
